Forget the distinction between long and int.
If we do int("4344554545454545") the result is 4344554545454545L, so the
same as long. There's no need do distinguish the two cases in our code
then. Moreover, there's really no way to know what will be mapped as int
and what will be mapped as long: python 'int' may be 16bit, or 32, or
64, or... depending on the machine. It was wrong for 32 bits now,
DevULong does not fit in a 32bit int. It is also better for future
portability to 3.0, where long will not exist anymore.