Default tango attr format %6.2f -> I DONT LIKE IT
I leave it for 'float' types, but for 'int's '%d' is a better default.
The problem is that there's no way to know if '%6.2f' is the default
value or if it has been set this way on purpose. Hopefully it will be
modified in TangoC++ someday and we could get rid of this commit.