TAttribute: raise exceptions from cache too
If we are not subscribed but read the value each time, until now
it was checking the timestamp only if the last lecture had worked,
but not if it had failed. But Tango itself is checking this, and will
throw an exception complaining about retryiing too soon and won't try
to read anyway. So there's no point, it works the same if theres an
exception or a successful read now.