PyQt SIGNAL disconnected problem ugly fix
In an ideal world the signature of this method should be
(evt_src, evt_type, evt_value). However there's a bug in PyQt:
If a signal is disconnected between the moment it is emitted and
the moment the slot is called, then the slot is called without
parameters (!?). We added this default values to detect if
this is the case without printing an error message each time.
If this gets fixed, we should remove this line.